+++ src/main/java/com/devsong/server/post/repository/CommentRepository.java
@@ -1,13 +1,11 @@
-package com.devsong.server.comment.repository;
+package com.devsong.server.post.repository;
 
-import com.devsong.server.comment.entity.Comment;
+import com.devsong.server.post.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDateTime;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
-    Comment findByCreatedAt(LocalDateTime createdAt);
 }
