+++ src/main/java/com/devsong/server/post/repository/PostLikeRepository.java
@@ -1,13 +1,11 @@
-package com.devsong.server.postlike.repository;
+package com.devsong.server.post.repository;
 
 import com.devsong.server.post.entity.PostLike;
-import com.devsong.server.postlike.entity.PostLike1;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostLikeRepository extends JpaRepository<PostLike1, Long>  {
-    PostLike1 findByPostId(Long postId);
+public interface PostLikeRepository extends JpaRepository<PostLike, Long>  {
     boolean existsByUserIdAndPostId(Long userId, Long postId);
     void deleteByUserIdAndPostId(Long userId, Long postId);
 }
