--- src/main/java/com/devsong/server/postapply/entity/PostApply.java
@@ -1,21 +1,0 @@
-package com.devsong.server.postapply.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostApply {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; //pk
-
-    @Column(nullable = false)
-    private Long userId; //유저
-
-    @Column(nullable = false)
-    private Long postId; //게시글
-}
