--- src/main/java/com/devsong/server/postapply/repository/PostApplyRepository.java
@@ -1,8 +1,0 @@
-package com.devsong.server.postapply.repository;
-
-import com.devsong.server.postapply.entity.PostApply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostApplyRepository extends JpaRepository<PostApply, Long> {
-    PostApply findByPostId(Long postId);
-}
