--- src/main/java/com/devsong/server/postlike/entity/PostLike1.java
@@ -1,21 +1,0 @@
-package com.devsong.server.postlike.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostLike1 {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; //pk
-
-    @Column(nullable = false)
-    private Long userId; //유저
-
-    @Column(nullable = false)
-    private Long postId; //게시글
-}
